tests/wpt/tests/html/editing/dnd/images/008.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="../resources/test-helper.js"></script>
<head>
  <title>dataURL image drag and drop between iframes</title>
  <style type="text/css">
    iframe {
      width: 160px;
      height: 160px;
      border-style: none;
    }
  </style>
</head>
<body>
  <p><iframe src="helper-drag-me-data-url-image.xhtml" id="drag-iframe">XHTML with image</iframe></p>
  <p>Drag circle above to the box below. It should be copied to the box once you drop it there.</p>
  <p><iframe src="helper-drop-image-here.xhtml" id="drop-iframe">XHTML with image</iframe></p>
  <script>
    async function test() {
      await new Promise(loaded => window.addEventListener("load", loaded));
      const dragIframe = document.getElementById('drag-iframe');
      const dropIframe = document.getElementById('drop-iframe');
      const div = frames[1].document.querySelector('div');

      function onDropCallBack(event) {
        const img = frames[0].document.querySelector('img');
        assert_equals(div.querySelector('img').src, img.src);
        return true;
      }
      dragDropTest(dragIframe, div, onDropCallBack,
        'Dragging a dataURL image within an iframe to another iframe should copy it there', dropIframe);
    }
    test();
  </script>
</body>
</html>
